#78 Let IO filter provide message minimum size

+++ io/src/main/java/org/jruyi/io/channel/IChannelService.java
@@ -24,7 +24,7 @@ public interface IChannelService {
 
 	public IBufferFactory getBufferFactory();
 
-	public int throttle();
+	public long throttle();
 
 	public IFilter<?, ?>[] getFilterChain();
