+++ io/src/main/java/org/jruyi/io/udpclient/UdpClient.java
@@ -79,8 +79,8 @@ public final class UdpClient extends Service implements IChannelService,
 	}
 
 	@Override
-	public int throttle() {
-		return 0;
+	public long throttle() {
+		return 0L;
 	}
 
 	@Override
